+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Changed:
 - Material icons in bottom nav
+- Material icons in audio recorder
 
 ## [0.8.313] - 2023-04-11
 ### Changed: